requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
adafruit-circuitpython-motorkit>=1.6.0
vosk>=0.3.45
sounddevice>=0.4.6
python-vlc>=3.0.18
